Added authentication blueprint and added login route to it
