+++ api/templates/base.html
@@ -19,7 +19,7 @@
       <nav>
         <a href="{{ url_for('home') }}">Home</a>
         <a href="{{ url_for('info.about') }}">About</a>
-        <a href="{{ url_for('login') }}">Login</a>
+        <a href="{{ url_for('auth.login') }}">Login</a>
       </nav>
     </header>
     {% endblock header %}
@@ -37,7 +37,7 @@
       <div class="hero">
         <h1>Welcome to QUIZIVERSE</h1>
         <p>Start your journey into a universe of quizzes.</b></p>
-        <a href="{{ url_for('login') }}"><button class="get-started-btn">Get Started</button></a>
+        <a href="{{ url_for('auth.login') }}"><button class="get-started-btn">Get Started</button></a>
       </div>
       <section class="features">
         <div class="feature">
