+++ api/templates/register.html
@@ -23,7 +23,7 @@
         <label>Password: <input type="password" name="password" autocomplete="off" required/></label><br>
         <label>Confirm Password: <input type="password" name="confirm_password" autocomplete="off" required /></label><br>
         <button type="submit">Register</button>
-        <p>Already have an account? <a href="{{ url_for('login') }}" class="login-link">Login</a></p>
+        <p>Already have an account? <a href="{{ url_for('auth.login') }}" class="login-link">Login</a></p>
     </form>
   </div>
 </section>
